move Menu component
